requirements.txt
streamlit
requests
openai
googlesearch-python
google-generativeai
python-dotenv
